requirements.txt.txt
folium==0.17.0
haversine==2.8.1
matplotlib-inline==0.1.7
numpy==1.24.4
pandas==2.0.3
pillow==10.4.0
plotly==5.23.0
streamlit==1.37.0
streamlit_folium==0.22.0
